+++ Sources/AMQPProtocol/Constants.swift
@@ -249,6 +249,13 @@ public enum ExchangeTypeName: String, Sendable {
   case fanout = "fanout"
   case topic = "topic"
   case headers = "headers"
+  // In RabbitMQ core since 4.3.0
+  case modulusHash = "x-modulus-hash"
+  // In RabbitMQ core since 4.2.0
+  case localRandom = "x-local-random"
+  // Provided by commonly used plugins
+  case consistentHash = "x-consistent-hash"
+  case random = "x-random"
 }
 
 // MARK: - Delivery Mode